pure refactor into common Pmtiles class/ deserializeDirectory -> directoryFromBytes, etc
