+++ planetiler-benchmarks/src/main/java/com/onthegomap/planetiler/benchmarks/BenchmarkPmtiles.java
@@ -2,7 +2,7 @@ package com.onthegomap.planetiler.benchmarks;
 
 import static io.prometheus.client.Collector.NANOSECONDS_PER_SECOND;
 
-import com.onthegomap.planetiler.pmtiles.WriteablePmtiles;
+import com.onthegomap.planetiler.pmtiles.Pmtiles;
 import com.onthegomap.planetiler.stats.Timer;
 import com.onthegomap.planetiler.util.Format;
 import java.io.IOException;
@@ -19,18 +19,18 @@ public class BenchmarkPmtiles {
 
     for (int i = 0; i < 3; i++) {
 
-      var entries = new ArrayList<WriteablePmtiles.Entry>();
+      var entries = new ArrayList<Pmtiles.Entry>();
 
       long offset = 0;
       for (int j = 0; j < num; j++) {
         int len = 200 + random.nextInt(64000);
-        entries.add(new WriteablePmtiles.Entry(j, offset, len, 1));
+        entries.add(new Pmtiles.Entry(j, offset, len, 1));
         offset += len;
       }
 
       var timer = Timer.start();
 
-      var result = WriteablePmtiles.deserializeDirectory(WriteablePmtiles.serializeDirectory(entries));
+      var result = Pmtiles.directoryFromBytes(Pmtiles.directoryToBytes(entries));
       assert (result.size() == entries.size());
 
       System.err.println(
